requirement.txt
beautifulsoup4
feedparser
pymongo
newspaper3k
spacy
gensim